Split image generation request context
